* Fixed warnings reported by GCC on MacOSX
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8555 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/attribute.cpp
@@ -94,7 +94,7 @@ void Attribute::moveValue(
 // ------------------------------------
 void Attribute::configValues(
     const vector< string > &/*curDomain*/,
-    map< int, int  > &/*origToCurDomain*/ )
+    map< size_t, size_t  > &/*origToCurDomain*/ )
 // ------------------------------------
 {}
